refactor: switch remaining blocks off lib-franklin
