+++ blocks/author-box/author-box.js
@@ -1,6 +1,6 @@
 import {
   getMetadata,
-} from '../../scripts/lib-franklin.js';
+} from '../../scripts/aem.js';
 
 import {
   createTag,
