+++ blocks/copy/copy.js
@@ -1,4 +1,4 @@
-import { readBlockConfig } from '../../scripts/lib-franklin.js';
+import { readBlockConfig } from '../../scripts/aem.js';
 
 export default function decorate(block) {
   const config = readBlockConfig(block);
