+++ blocks/service-status/service-status.js
@@ -1,4 +1,4 @@
-import { loadScript } from '../../scripts/lib-franklin.js';
+import { loadScript } from '../../scripts/aem.js';
 import createTag from '../../utils/tag.js';
 
 async function loadWidget(el) {
